+++ src/test/groovy/de/thomaskrille/dropwizard_template_config/BundleCreationSpec.groovy
@@ -11,13 +11,19 @@ class BundleCreationSpec extends Specification {
 
         then:
         bundle.charset == Charsets.UTF_8
+        bundle.includePath == Optional.absent()
     }
 
     def 'a specific configuration can be applied'() {
         when:
-        def bundle = new TemplateConfigBundle(new TemplateConfigBundleConfiguration().charset(Charsets.US_ASCII))
+        def bundle = new TemplateConfigBundle(
+                new TemplateConfigBundleConfiguration()
+                        .charset(Charsets.US_ASCII)
+                        .includePath('includePath')
+        )
 
         then:
         bundle.charset == Charsets.US_ASCII
+        bundle.includePath.get() == 'includePath'
     }
 }