+++ src/test/groovy/de/thomaskrille/dropwizard_template_config/EnvironmentVariablesSpec.groovy
@@ -1,7 +1,7 @@
 package de.thomaskrille.dropwizard_template_config
 
 import com.google.common.base.Charsets
-import com.google.common.io.CharStreams
+import com.google.common.base.Optional
 import org.apache.commons.io.IOUtils
 import spock.lang.Specification
 
@@ -16,7 +16,7 @@ class EnvironmentVariablesSpec extends Specification {
             new TemplateConfigurationSourceProvider(new TestConfigSourceProvider(),
                     environmentProvider,
                     new DefaultSystemPropertiesProvider(),
-                    Charsets.UTF_8)
+                    Charsets.UTF_8, Optional.absent())
 
     def 'replacing an environment variable works'() throws Exception {
         given: